Solution to Check Greater Than

+++ Check Greater Than@@67882a731b4392d8fc1577d6/code.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 int main() {
-    long int x,y;
+    int x,y;
     scanf("%d %d",&x,&y);
     if (x>y){
         printf("True");
